add existing tapNextRelease code to Examples.elm
I had tried to reimplement TNR with pattern matching on a new Event
type, but I never got it compiling so it will continue to exist as a
comment. In the future, I could generalize the implementation and move
it to Layer.elm as "mkTapNextRelease" and then create a test layer in
examples that invokes mkTapNextRelease.